separate navigation into components
