+++ src/IbdTracker.Frontend/src/components/dashboard/navigation/Navigation.svelte
@@ -1,0 +1,17 @@
+<script lang="ts">
+    import Headbar from "./Headbar.svelte";
+    import Menu from "./Menu.svelte";
+
+    let menuOpened: Boolean = false;
+
+    function openMenu(): void {
+        menuOpened = true;
+    }
+
+    function closeMenu(): void {
+        menuOpened = false;
+    }
+</script>
+
+<Headbar on:click={openMenu} />
+<Menu {menuOpened} on:click={closeMenu} />
